chore(OpenQueryButtons): render YQLButton from the component [YTFRONT-4463]
