+++ packages/ui/src/ui/pages/navigation/content/Document/DocumentBody.scss
@@ -1,0 +1,5 @@
+.yt-document-body {
+    &__yql {
+        margin-right: 10px;
+    }
+}
