+++ packages/ui/src/ui/pages/navigation/content/Table/TableOverview/CreateQueryFromTable.tsx
@@ -3,11 +3,11 @@ import {useSelector} from 'react-redux';
 
 import {getPath} from '../../../../../store/selectors/navigation';
 import {getCluster} from '../../../../../store/selectors/global';
-import {OpenQueryButton} from '../../../../../containers/OpenQueryButton/OpenQueryButton';
+import {OpenQueryButtons} from '../../../../../containers/OpenQueryButtons/OpenQueryButtons';
 
 export function CreateQueryFromTable({className}: {className: string}) {
     const cluster = useSelector(getCluster);
     const path = useSelector(getPath);
 
-    return <OpenQueryButton className={className} path={path} cluster={cluster} />;
+    return <OpenQueryButtons className={className} path={path} cluster={cluster} />;
 }
