+++ packages/ui/src/ui/pages/navigation/content/Table/TableOverview/TableOverview.scss
@@ -108,8 +108,6 @@ $overview-height: 50px;
     &__yql {
         margin-left: 10px;
         margin-right: 10px;
-
-        display: flex;
     }
 
     &__jupyter {
